fix(stream): fix go-jsonnet type compatibility in OTLP flatten code
Convert all uint32/uint64/typed-slice fields to go-jsonnet-compatible
types (int64/[]any) to prevent "Not a json type" errors at record write time.
Also add empty-batch source validation so wrong-secret requests get 404
instead of a free 200 on empty OTLP payloads.

+++ internal/pkg/service/stream/source/type/otlpsource/flatten_logs.go
@@ -41,10 +41,10 @@ func flattenLogRecord(
 	rec := orderedmap.New()
 	rec.Set("timestamp", formatTimestamp(lr.Timestamp()))
 	rec.Set("observed_timestamp", formatTimestamp(lr.ObservedTimestamp()))
-	rec.Set("severity_number", int32(lr.SeverityNumber()))
+	rec.Set("severity_number", int64(lr.SeverityNumber()))
 	rec.Set("severity_text", lr.SeverityText())
 	rec.Set("body", anyValueToInterface(lr.Body()))
-	rec.Set("flags", uint32(lr.Flags()))
+	rec.Set("flags", int64(lr.Flags()))
 
 	// trace_id and span_id are omitted (not "") when empty — the Path column
 	// will fall back to its defaultValue. Emitting "" would mask the absence.
